start admin

+++ WebBanDT/Models/ViewModels/CartItemViewModel.cs
@@ -3,6 +3,6 @@
 	public class CartItemViewModel
 	{
 		public List<CartItemModel> CartItems { get; set; }
-		public decimal GrandTotal {  get; set; }
+		public double GrandTotal {  get; set; }
 	}
 }
